Continuing old exercise.
Experimenting with Niels Hazelaars' suggestions

Updating files on remote with local files

It seems the remote mirrors an old commit to the *previous* exercise
from 26 oct '21. I suspect during the big bitlab incident from a month
ago, this repo got "reinitialised" and standard behaviour is to fork the
previous exercise when the new one continues on the same subject.